refactor: 💡 merge `entityTypes.ts` into `entities/types.ts`
